Allow editing of params at any time
Currently it is only possible to edit command line parameters in dev
mode after the application has exited. This means that if you have long
running applications (or accidently hang your app) you can't restart
with different parameters.
